changed code (again) for integer

+++ _integer.c
@@ -10,8 +10,8 @@
 int _integer(va_list args)
 {
 	int num = va_arg(args, int);
-	int count = 0;
-	int divisor;
+	long int count = 0;
+	long int divisor;
 	int multiplication = 1;
 
 	if (num < 0)
